+++ run_server.sh
@@ -6,4 +6,4 @@ else
 fi
 
 PYRAMID_RELOAD_TEMPLATES=1 ${GUNICORN_BIN} -b 127.0.0.1:7001 -w 1 -t 600 \
-    -c ichnaea.webapp.settings ichnaea.webapp.app:wsgi_app
+    -c python:ichnaea.webapp.settings ichnaea.webapp.app:wsgi_app